fixed the student update and delete

+++ app/api/student-registration/[id]/route.ts
@@ -1,4 +1,5 @@
 import { prisma } from "@/lib/prisma";
+import { ApiError } from "@/utils/ApiError";
 import { apiHandler } from "@/utils/apiHandler";
 import { ApiResponse } from "@/utils/ApiResponse";
 import { NextResponse } from "next/server";
@@ -6,6 +7,8 @@ import { NextResponse } from "next/server";
 export const PUT = apiHandler(async (req: Request, context: any) => {
   const { id } = context.params;
   const body = await req.json();
+
+  if (!id) throw new ApiError(400, "student ID is required");
 
   delete body.id;
   delete body.createdAt;
@@ -24,8 +27,27 @@ export const PUT = apiHandler(async (req: Request, context: any) => {
 
 export const DELETE = apiHandler(async (_req: Request, context: any) => {
   const { id } = context.params;
+  if (!id) throw new ApiError(400, "student ID is required");
   await prisma.studentRegistration.delete({ where: { id } });
   return NextResponse.json(
     new ApiResponse(200, null, "Student deleted successfully")
+  );
+});
+
+export const GET = apiHandler(async (req: Request, context: any) => {
+  const { id } = context.params;
+
+  if (!id) throw new ApiError(400, "student ID is required");
+
+  const student = await prisma.studentRegistration.findUnique({
+    where: { id },
+  });
+
+  if (!student) {
+    throw new ApiError(404, `No student found with this id ${id}`);
+  }
+
+  return NextResponse.json(
+    new ApiResponse(200, student, "user fetched successfully")
   );
 });
